Allow configuring NTP pools and servers for LBMs (#411)
* Allow configuring NTP pools and servers for LBMs

* Support new config in chart

* Format and improve linter settings

* go mod tidy

* Remove chronyd depencency in favor of restart

It seems desirable to start chrony as soon as possible even if it doesn't have its
final config yet. If the clock is off by too much, chrony should perform a step instead
of slewing the clock. It's better to do this quickly after boot before time-sensitive
processes might notice.

* Replace deprecated `initstepslew` directive

* Comment on not using cloud-init's NTP module

* earthly +generate